# canonifyurls in srcset